requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.7
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
Markdown==3.6
pycparser==2.22
PyJWT==2.8.0
sqlparse==0.5.0
